Add separate `fuzz.rs`
